requirements.txt
"""
requirements.txt - Python dependencies for DrugBot
"""

# Core ML Libraries
torch>=1.9.0
torchvision
torchaudio

# Transformers and NLP
transformers>=4.20.0
unsloth
bitsandbytes
accelerate
peft
trl
sentencepiece

# Vector Search and Embeddings
faiss-cpu
langchain-community

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Hugging Face
huggingface_hub

# Progress bars
tqdm

# Web scraping (optional)
requests
beautifulsoup4

# Jupyter (for notebook development)
jupyter
ipython

# Testing
pytest
pytest-cov

# Code quality
black
flake8
isort

# Documentation
sphinx
sphinx-rtd-theme